Refactor : 그룹 드롭다운 로직 GroupDropdownSelector 컴포넌트로 분리
